requirements.txt
SQLAlchemy==2.0.22
Alembic==1.12.0
psycopg2-binary==2.9.9
fastpi==0.104.0
passlib==1.7.4
uvicorn==0.23.2
pyjwt==2.8.0